+++ Changelog.md
@@ -1,6 +1,7 @@
 --------------------------Changelog for v1.9.2 of SecurityCraft--------------------------
 
 - Fix: Breaking the block another block with modules is attached to does not drop the modules
+- Fix: Breaking the block a Key Panel is attached to, does not drop the Key Panel
 
 --------------------------Changelog for v1.9.1 of SecurityCraft--------------------------